Initialize config (host, port) once per VCL instead of once every call
Because the VCC specifies PRIV_CALL for vmod_send() and vmod_call(), the
vmod_priv argument is NULL for every call to these functions. This causes
the config struct to be re-allocated for every call.

This patch changes PRIV_CALL to PRIV_VCL, making the vmod_priv argument shared
between all calls to the module's functions within a given instance of a VCL.

The allocation of the config struct is also moved from redis_common() to
init_function().

Signed-off-by: Noah Williamsson <noah.williamsson@gmail.com>